refactor(step2): Step2를 타입스크립트로 리팩터링 (#365)
- ErrorReportModalComponent.tsx
